StoreWAL: handle case when write cache becomes inconsistent if commit dies. Fix StoreWALTest.crash_with_interrupt

+++ src/test/java/org/mapdb/StoreAppendTest.java
@@ -43,7 +43,19 @@ public class StoreAppendTest<E extends StoreAppend> extends EngineTest<E>{
 
     @After
     public void deleteFile(){
+        if(e!=null && !e.isClosed()){
+            e.close();
+            e = null;
+        }
+        if(f==null)
+            return;
+
         f.delete();
+        String name = f.getName();
+        for(File f2:f.getParentFile().listFiles()){
+            if(f2.getName().startsWith(name))
+                f2.delete();
+        }
     }
 
     @Override
